# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: go-orca API
  description: |
    go-orca is a self-hosted AI workflow orchestration server. It drives
    multi-phase LLM pipelines (Director → Project Manager → Architect →
    Implementer → QA → Finalizer) and delivers structured artifacts via
    GitHub PR, markdown export, webhook, or artifact bundle.

    ## Authentication
    go-orca does not include built-in authentication. Secure the API with a
    reverse proxy (nginx, Caddy, etc.) or network-level controls.

    ## Multi-tenancy
    All workflow and scope operations are scoped to a tenant. Pass
    `X-Tenant-ID` and `X-Scope-ID` headers to select a specific tenant and
    scope. When omitted, the server default tenant and scope are used.

    ## SSE Streaming
    `GET /workflows/{id}/stream` returns a `text/event-stream` response.
    Set `proxy_buffering off` in your reverse proxy to receive events in
    real time. When Redpanda is enabled (`streaming.enabled`), live events
    are delivered from the workflow topic via an in-process hub after a
    database snapshot; use query `source=auto|redpanda|database` and
    `GET /streaming` for capabilities. See docs/streaming.md.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  parameters:
    TenantID:
      name: X-Tenant-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: >
        Tenant context for the request. Falls back to the server default
        tenant when omitted.
    ScopeID:
      name: X-Scope-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: >
        Scope context for the request. Falls back to the server default
        scope when omitted.
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required:
        - error

  # ─── Reusable responses ──────────────────────────────────────────────────────
  responses:
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "workflow not found"
...
    BadRequest:
      description: Invalid request body or missing required field.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "request field is required"
...
paths:
...
  /api/v1/persona-runs:
    post:
      summary: Run a single persona ad-hoc
      operationId: createPersonaRun
      tags:
        - workflows
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ScopeID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [persona]
              properties:
                persona: { type: string }
                request: { type: string }
                mode: { type: string }
                provider: { type: string }
                model: { type: string }
                workflow_id: { type: string, format: uuid }
                persist: { type: boolean }
                tools_scope: { type: string, enum: [full, mcp_agent, none] }
                timeout_sec: { type: integer }
                workspace_path: { type: string }
      responses:
        "200":
          description: Persona run result
        "400":
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/workflow-templates/{id}:
    get:
      summary: Get workflow template by id
      operationId: getWorkflowTemplate
      tags:
        - workflows
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Template definition
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Workflows ───────────────────────────────────────────────────────────────
...